+++ client/src/Providers/spItemsRepository.ts
@@ -499,7 +499,10 @@ export class ItemsRepository implements CompletionItemProvider, Disposable {
     const tokensBuilder = new SemanticTokensBuilder(SP_LEGENDS);
     let allItems: SPItem[] = this.getAllItems(document.uri.toString());
     for (let item of allItems) {
-      if (item.kind === CompletionItemKind.Constant) {
+      if (
+        item.kind === CompletionItemKind.Constant ||
+        item.kind === CompletionItemKind.EnumMember
+      ) {
         for (let call of item.calls) {
           if (call.uri.fsPath === document.uri.fsPath) {
             tokensBuilder.push(call.range, "variable", ["readonly"]);