+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Complete refactoring of the GoToDefinition system, which now updates the location of an item every parsing.
 - Added support for signatures help in multi-line function calls.
 - Added escaped characters highlighting.
-- Added dynamic syntax highlighting for defines.
+- Added dynamic syntax highlighting for defines and enum Members.
 
 ### Fixed